requirements.txt
discord.py
pandas
gspread
oauth2client
opencv-python-headless
google-cloud-vision
thefuzz